+++ Assets/__Scripts/Settings/Settings.cs
@@ -21,7 +21,7 @@ public class Settings {
     public bool OSC_Enabled = false;
     public string OSC_IP = "127.0.0.1";
     public string OSC_Port = "8080";
-    public int EditorScale = 4;
+    public float EditorScale = 4;
     public int ChunkDistance = 5;
     public int AutoSaveInterval = 5;
     //public int InitialLoadBatchSize = 100;
